Reorganized the files, made subdirectories (core, managers, utilities, ui) and amended the makefile to fit the new structure, fixed the headers/include
