+++ src/core/client.h
