+++ src/core/material.cpp
