+++ src/core/reports.cpp
