+++ src/core/reports.h
@@ -6,7 +6,7 @@
 #include <memory>
 #include "order.h"
 #include "material.h"
-#include "clocks.h"
+#include "utilities/clocks.h"
 
 class Report {
     public:
