+++ src/managers/clientmanager.cpp
