+++ src/managers/materialmanager.cpp
